services/payment_service/requirements.txt
Django==4.2.0
djangorestframework==3.14.0
mysqlclient==2.1.1
django-cors-headers==4.0.0
pika==1.3.1
requests==2.28.2
gunicorn==20.1.0
Pillow>=9.0.0
stripe==8.0.0
python-dotenv